refac: move gamemode into own file
