+++ src/net/mod.rs
@@ -1,6 +1,8 @@
 pub mod connection;
 pub mod encoding;
 mod packet;
+mod game_mode;
 pub mod udp_conn;
 
 pub use packet::*;
+pub use game_mode::*;
